+++ R/gee.R
@@ -241,7 +241,7 @@ GEEModuleLinear <- function(input, output, session, data, data_label, data_varSt
   out <- reactive({
     data.regress <- data()
     label.regress <- data_label()
-    id <- id.gee()
+    idgee_Plz_Noduplicate <- id.gee()
     if(input$regressUI_subcheck == T){
       req(input$subvar_regress)
       if (input$subvar_regress %in% vlist()$factor_vars){
@@ -272,9 +272,9 @@ GEEModuleLinear <- function(input, output, session, data, data_label, data_varSt
     )
 
     nomiss <- stats::complete.cases(data.regress[, c(y, xs), with = F])
-    res.gee <- geepack::geeglm(form, data = data.regress[nomiss, ], family = "gaussian", id = eval(id), corstr = "exchangeable")
+    res.gee <- geepack::geeglm(form, data = data.regress[nomiss, ], family = "gaussian", id = get(idgee_Plz_Noduplicate), corstr = "exchangeable")
     info.gee <- jstable::geeglm.display(res.gee, decimal = input$decimal)
-    info.gee$caption = gsub("id", id, info.gee$caption)
+    info.gee$caption <- gsub("idgee_Plz_Noduplicate", idgee_Plz_Noduplicate, info.gee$caption)
     ltb.gee <- jstable::LabeljsGeeglm(info.gee, ref = label.regress)
     out.tb <- rbind(ltb.gee$table, ltb.gee$metric)
     cap.gee <- ltb.gee$caption
@@ -481,7 +481,7 @@ GEEModuleLogistic <- function(input, output, session, data, data_label, data_var
   out <- reactive({
     data.logistic <- data()
     label.regress <- data_label()
-    id <- id.gee()
+    idgee_Plz_Noduplicate <- id.gee()
     if(input$regressUI_subcheck == T){
       req(input$subvar_regress)
       if (input$subvar_regress %in% vlist()$factor_vars){
@@ -514,9 +514,9 @@ GEEModuleLogistic <- function(input, output, session, data, data_label, data_var
     )
 
     nomiss <- stats::complete.cases(data.logistic[, c(y, xs), with = F])
-    res.gee <- geepack::geeglm(form, data = data.logistic[nomiss, ], family = "binomial", id = eval(id), corstr = "exchangeable")
+    res.gee <- geepack::geeglm(form, data = data.logistic[nomiss, ], family = "binomial", id = get(idgee_Plz_Noduplicate), corstr = "exchangeable")
     info.gee <- jstable::geeglm.display(res.gee, decimal = input$decimal)
-    info.gee$caption = gsub("id", id, info.gee$caption)
+    info.gee$caption = gsub("idgee_Plz_Noduplicate", idgee_Plz_Noduplicate, info.gee$caption)
     ltb.gee <- jstable::LabeljsGeeglm(info.gee, ref = label.regress)
     out.tb <- rbind(ltb.gee$table, ltb.gee$metric)
     cap.gee <- ltb.gee$caption